+++ core/src/main/kotlin/su/plo/voice/discs/item/DiscHelper.kt
@@ -1,0 +1,10 @@
+package su.plo.voice.discs.item
+
+import org.bukkit.inventory.ItemStack
+
+/**
+ * @since Minecraft 1.21.3
+ */
+interface DiscHelper {
+    fun showSongTooltip(item: ItemStack, show: Boolean)
+}